+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - maxSelected option
 - searchDebounce option
+- strings option
 - eslint config
 - Message when not items selected
 - Selector loading spinner